+++ lib/karafka/pro/routing/features/throttling/topic.rb
@@ -18,17 +18,8 @@ module Karafka
         class Throttling < Base
           # Topic throttling API extensions
           module Topic
-            # @param factory [nil, Class] nil if we want to use the default throttler
-            #   with the `limit` and `interval` provided via the standard API or a class from
-            #   which we may build custom throttlers.
-            #   custom throttling class that we can use to provide custom throttling capabilities
             # @param limit [Integer] max messages to process in an time interval
             # @param interval [Integer] time interval for processing
-            # @note We allow for direct definition of `limit` and `interval` here despite them
-            #   being used only by the default throttler, because we want to provide a simple and
-            #   out-of-the-box API that will not force users to define their own throttlers for
-            #   simple cases. The `factory` overwrite is suppose to be for more advanced
-            #   users.
             def throttling(
               limit: Float::INFINITY,
               interval: 60_000
@@ -41,8 +32,8 @@ module Karafka
                   interval: interval
                 )
 
-                # If someone defined throttling setupp, we need to create appropriate filter
-                # for it
+                # If someone defined throttling setup, we need to create appropriate filter for it
+                # and inject it via filtering feature
                 if config.active?
                   factory = -> { Pro::Processing::Filters::Throttler.new(limit, interval) }
                   filter(factory)
@@ -50,6 +41,13 @@ module Karafka
 
                 config
               end
+            end
+
+            # Just an alias for nice API
+            #
+            # @param args [Array] Anything `#throttling` accepts
+            def throttle(**args)
+              throttling(**args)
             end
 
             # @return [Boolean] is a given job throttled