+++ spec/lib/karafka/pro/routing/features/expiring/contract_spec.rb
@@ -1,0 +1,36 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  subject(:check) { described_class.new.call(config) }
+
+  let(:config) do
+    {
+      expiring: {
+        active: true,
+        ttl: 5
+      }
+    }
+  end
+
+  context 'when config is valid' do
+    it { expect(check).to be_success }
+  end
+
+  context 'when active flag is not boolean' do
+    before { config[:expiring][:active] = rand }
+
+    it { expect(check).not_to be_success }
+  end
+
+  context 'when ttl is zero' do
+    before { config[:expiring][:ttl] = rand }
+
+    it { expect(check).not_to be_success }
+  end
+
+  context 'when ttl is not an integer' do
+    before { config[:expiring][:ttl] = 'test' }
+
+    it { expect(check).not_to be_success }
+  end
+end