+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Karafka framework changelog
 
 ## 2.0.39 (Unreleased)
-- **[Feature]** Provide ability to apply filters after messages are polled and before enqueued.
 - **[Feature]** Provide ability to throttle/limit number of messages processed in a time unit (#1203)
 - **[Feature]** Provide Delayed Topics (#1000)
 - **[Feature]** Provide ability to expire messages (expiring topics)
+- **[Feature]** Provide ability to apply filters after messages are polled and before enqueued. This is a generic filter API for any usage.
 - [Improvement] Track active_job_id in instrumentation (#1372)
 - [Improvement] Introduce new housekeeping job type called `Idle` for non-consumption execution flows.
 - [Improvement] Change how a manual offset management works with Long-Running Jobs. Use the last message offset to move forward instead of relying on the last message marked as consumed for a scenario where no message is marked.